+++ client/src/primitive/triangle/builder.rs
@@ -1,8 +1,8 @@
 use super::Triangle;
 use crate::primitive::point::Point;
 
-pub fn triangle(input: [[i32; 2]; 3]) -> Triangle {
-    let mut points = [Point { x: 0, y: 0 }; 3];
+pub fn triangle(input: [[f32; 2]; 3]) -> Triangle {
+    let mut points = [Point { x: 0.0, y: 0.0 }; 3];
     for (point, i) in points.iter_mut().zip(input) {
         point.x = i[0];
         point.y = i[1];
